select cohort helper
